+++ test/test_helper.rb
@@ -3,7 +3,4 @@ Bundler.require(:default)
 require "minitest/autorun"
 require "minitest/pride"
 
-ENV["PGSLICE_ENV"] = "test"
-
-$url = ENV["PGSLICE_URL"] || "postgres:///pgslice_test"
-$conn = PG::Connection.new($url)
+PgSlice::CLI.exit_on_failure = false